==> 3. Ensemble.ipynb <==
import pandas as pd

# Read the first CSV file
df1 = pd.read_csv("merged_file_transformer_12heads.csv")

# Read the second CSV file
df2 = pd.read_csv("..\merged_file.csv")

# Convert DataFrame values to numeric type
df1.iloc[:,1:] = (df1.iloc[:,1:] + df2.iloc[:,1:]) / 2

# Write the averages to a new CSV file
df1.to_csv("ensemble_average.csv", index=False)


import pandas as pd
import numpy as np

# Read the first CSV file
df1 = pd.read_csv("merged_file_transformer_12heads.csv", index_col="id")

# Read the second CSV file
df2 = pd.read_csv("..\merged_file.csv", index_col="id")

# Take the element-wise maximum between df1 and df2
max_values = np.maximum(df1, df2)

# Write the updated DataFrame to a new CSV file
max_values.to_csv("ensemble_max.csv")


import pandas as pd
import numpy as np

# Read the first CSV file
df1 = pd.read_csv("merged_file_transformer_12heads.csv", index_col="id")

# Read the second CSV file
df2 = pd.read_csv("..\merged_file.csv", index_col="id")

# Take the element-wise maximum between df1 and df2
max_values = np.minimum(df1, df2)

# Write the updated DataFrame to a new CSV file
max_values.to_csv("ensemble_min.csv")
