==> 4. Another_Approach_Transformer.ipynb <==
import numpy as np
import pandas as pd

timesteps = 14
startDay = 350

dt = pd.read_csv("m5-forecasting-accuracy/sales_train_evaluation.csv")
dt.head(3)

dt.info()

#To reduce memory usage
def downcast_dtypes(df):
    float_cols = [c for c in df if df[c].dtype == "float64"]
    int_cols = [c for c in df if df[c].dtype in ["int64", "int32"]]
    df[float_cols] = df[float_cols].astype(np.float32)
    df[int_cols] = df[int_cols].astype(np.int16)
    return df

#Reduce memory usage and compare with the previous one to be sure
dt = downcast_dtypes(dt)

dt.info()

#Take the transpose so that we have one day for each row, and 30490 items' sales as columns
dt = dt.T    
dt.head(8)

#Remove id, item_id, dept_id, cat_id, store_id, state_id columns
dt = dt[6 + startDay:]
dt.head(5)

calendar = pd.read_csv("m5-forecasting-accuracy/calendar.csv")
calendar

#Create dataframe with zeros for 1969 days in the calendar
daysBeforeEvent = pd.DataFrame(np.zeros((1969,1)))

# "1" is assigned to the days before the event_name_1. Since "event_name_2" is rare, it was not added.
for x,y in calendar.iterrows():
    if((pd.isnull(calendar["event_name_1"][x])) == False):
           daysBeforeEvent[0][x-1] = 1 
            #if first day was an event this row will cause an exception because "x-1".
            #Since it is not i did not consider for now.

#"calendar" won't be used anymore. 
del calendar

#"daysBeforeEventTest" will be used as input for predicting (We will forecast the days 1913-1941)
daysBeforeEventTest = daysBeforeEvent[1941:1969]
#"daysBeforeEvent" will be used for training as a feature.
daysBeforeEvent = daysBeforeEvent[startDay:1941]

#Before concatanation with our main data "dt", indexes are made same and column name is changed to "oneDayBeforeEvent"
daysBeforeEvent.columns = ["oneDayBeforeEvent"]
daysBeforeEvent.index = dt.index

dt = pd.concat([dt, daysBeforeEvent], axis = 1)

dt.columns

dt

import pandas as pd
from sklearn.preprocessing import MinMaxScaler

# Assuming dt is a pandas DataFrame
dt.columns = dt.columns.astype(str)  # Convert column names to strings

sc = MinMaxScaler(feature_range=(0, 1))
dt_scaled = sc.fit_transform(dt)

X_train = []
y_train = []
for i in range(timesteps, 1941 - startDay):
    X_train.append(dt_scaled[i-timesteps:i])
    y_train.append(dt_scaled[i][0:30490]) 

del dt_scaled

#Convert to np array to be able to feed the LSTM model
X_train = np.array(X_train)
y_train = np.array(y_train)
print(X_train.shape)
print(y_train.shape)

X_train_token_ids = np.argmax(X_train, axis=-1)

import tensorflow as tf
from tensorflow.keras.models import Model
from tensorflow.keras.layers import Input, Dense, Dropout, LayerNormalization, MultiHeadAttention, Embedding
from tensorflow.keras.layers import Input, Embedding, Dense, Dropout, GlobalAveragePooling1D
from tensorflow.keras.optimizers import Adam

class PositionalEncoding(tf.keras.layers.Layer):
    def __init__(self, position, d_model):
        super(PositionalEncoding, self).__init__()
        self.pos_encoding = self.positional_encoding(position, d_model)

    def get_angles(self, position, i, d_model):
        angles = 1 / tf.pow(10000, (2 * (i // 2)) / tf.cast(d_model, tf.float32))
        return position * angles

    def positional_encoding(self, position, d_model):
        angle_rads = self.get_angles(position=tf.range(position, dtype=tf.float32)[:, tf.newaxis],
                                     i=tf.range(d_model, dtype=tf.float32)[tf.newaxis, :],
                                     d_model=d_model)
        # Apply sin to even indices in the array; 2i
        sines = tf.math.sin(angle_rads[:, 0::2])
        # Apply cos to odd indices in the array; 2i+1
        cosines = tf.math.cos(angle_rads[:, 1::2])
        pos_encoding = tf.concat([sines, cosines], axis=-1)[tf.newaxis, ...]
        return tf.cast(pos_encoding, tf.float32)

    def call(self, inputs):
        return inputs + self.pos_encoding[:, :tf.shape(inputs)[1], :]

class TransformerEncoderLayer(tf.keras.layers.Layer):
    def __init__(self, d_model, num_heads, dff, rate=0.1):
        super(TransformerEncoderLayer, self).__init__()

        self.mha = MultiHeadAttention(key_dim=d_model, num_heads=num_heads)
        self.ffn = tf.keras.Sequential([
          Dense(dff, activation='relu'),  # (batch_size, seq_len, dff)
          Dense(d_model)  # (batch_size, seq_len, d_model)
        ])

        self.layernorm1 = LayerNormalization(epsilon=1e-6)
        self.layernorm2 = LayerNormalization(epsilon=1e-6)

        self.dropout1 = Dropout(rate)
        self.dropout2 = Dropout(rate)

    def call(self, inputs, training):
        attn_output = self.mha(inputs, inputs, inputs)  # (batch_size, input_seq_len, d_model)
        attn_output = self.dropout1(attn_output, training=training)
        out1 = self.layernorm1(inputs + attn_output)

        ffn_output = self.ffn(out1)  # (batch_size, input_seq_len, d_model)
        ffn_output = self.dropout2(ffn_output, training=training)
        out2 = self.layernorm2(out1 + ffn_output)

        return out2

def build_transformer(num_layers, d_model, num_heads, dff, input_vocab_size, maximum_position_encoding, rate=0.1):
    inputs = Input(shape=(None,))
    x = Embedding(input_vocab_size, d_model)(inputs)
    x = PositionalEncoding(maximum_position_encoding, d_model)(x)

    for _ in range(num_layers):
        x = TransformerEncoderLayer(d_model, num_heads, dff, rate)(x)

    x = GlobalAveragePooling1D()(x)
    x = Dropout(rate)(x)
    outputs = Dense(1, activation='linear')(x)  # Assuming a regression task

    model = Model(inputs=inputs, outputs=outputs)
    return model

# Define model parameters
num_layers = 4
d_model = 128
num_heads = 4
dff = 512
input_vocab_size = 30500 
maximum_position_encoding = 1000  # This should cover the maximum sequence length

# Build the Transformer model
transformer = build_transformer(num_layers, d_model, num_heads, dff, input_vocab_size, maximum_position_encoding)

# Compile the model
transformer.compile(optimizer=Adam(1e-4), loss='mean_squared_error')

# Summary of the model to verify the architecture
transformer.summary()


transformer.fit(X_train_token_ids, y_train, epochs=32, batch_size=32)

inputs= dt[-timesteps:]
inputs = sc.transform(inputs)

timesteps = 14
input_vocab_size = 30500

# Simulating 'X_test' as a sequence of token IDs
import numpy as np
X_test = np.random.randint(0, input_vocab_size, size=(1, timesteps))

predictions = []

for j in range(28): 
    predicted_stock_price = transformer.predict(X_test[:, -timesteps:])
    new_predicted_token_id = np.random.randint(0, input_vocab_size, size=(1, 1))
    X_test = np.append(X_test, new_predicted_token_id, axis=1)
    predictions.append(new_predicted_token_id)

import time

predictions = [np.random.rand(30490) for _ in range(28)]

submission = pd.DataFrame(predictions).T  
submission = pd.concat([submission]*2, ignore_index=True)

sample_submission = pd.read_csv("m5-forecasting-accuracy/sample_submission.csv")
idColumn = sample_submission[["id"]]

idColumn = pd.concat([idColumn]*2, ignore_index=True)

submission.insert(0, "id", idColumn)

submission.columns = ["id"] + [f"F{i}" for i in range(1, 29)]

currentDateTime = time.strftime("%d%m%Y_%H%M%S")

submission

submission.to_csv("submission_transformer.csv", index=False)
